emr/requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.14.2
beepy==1.0.7
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.0
cmake==4.0.2
colorama==0.4.6
comm==0.2.3
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dlib==19.24.9
dnspython==2.7.0
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.104.1
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.1
flask-cors==5.0.1
flatbuffers==25.9.23
fonttools==4.61.1
fqdn==1.5.1
fsspec==2025.9.0
gast==0.7.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.17.3
idna==3.10
imaplib2==3.6
importlib-metadata==6.11.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.3
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.13.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.31
lark==1.3.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.1.0
narwhals==2.15.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.24.3
oauthlib==3.3.1
openai==1.109.1
opencv-python==4.8.1.78
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.3
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.5
Pillow==10.1.0
platformdirs==4.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.1.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.13.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.6
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.3.0
scipy==1.15.3
Send2Trash==1.8.3
simpleaudio==1.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.27.0
streamlit==1.52.2
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.14.1
toml==0.10.2
torch==2.1.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.35.0
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
uv==0.8.22
uvicorn==0.24.0
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.14.2
xlrd==2.0.1
XlsxWriter==3.2.3
zipp==3.23.0
zstandard==0.25.0
